+++ org.jacoco.core/src/org/jacoco/core/analysis/ILine.java
@@ -12,6 +12,8 @@
  *******************************************************************************/
 package org.jacoco.core.analysis;
 
+import java.math.BigInteger;
+
 /**
  * The instruction and branch coverage of a single source line is described by
  * this interface.
@@ -50,5 +52,5 @@ public interface ILine {
 	 *
 	 * @return execution count
 	 */
-	int getExecutionCount();
+	BigInteger getExecutionCount();
 }
