+++ org.jacoco.core/src/org/jacoco/core/internal/analysis/ClassAnalyzer.java
@@ -12,6 +12,7 @@
  *******************************************************************************/
 package org.jacoco.core.internal.analysis;
 
+import java.math.BigInteger;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -34,7 +35,7 @@ public class ClassAnalyzer extends ClassProbesVisitor
 		implements IFilterContext {
 
 	private final ClassCoverageImpl coverage;
-	private final int[] probes;
+	private final BigInteger[] probes;
 	private final StringPool stringPool;
 
 	private final Set<String> classAnnotations = new HashSet<String>();
@@ -55,8 +56,8 @@ public class ClassAnalyzer extends ClassProbesVisitor
 	 * @param stringPool
 	 *            shared pool to minimize the number of {@link String} instances
 	 */
-	public ClassAnalyzer(final ClassCoverageImpl coverage, final int[] probes,
-			final StringPool stringPool) {
+	public ClassAnalyzer(final ClassCoverageImpl coverage,
+			final BigInteger[] probes, final StringPool stringPool) {
 		this.coverage = coverage;
 		this.probes = probes;
 		this.stringPool = stringPool;
